requirements.txt
fastapi==0.115.5
pre-commit==4.0.1
psycopg2==2.9.10 
shapely==2.0.6
openpyxl==3.1.5
geojson==3.1.0
aiofiles==24.1.0